Modularized the application (directories/modules) layout
No code has been changed at this stage.
