+++ lib/pm_spotlight_daemon/resources/telescope-icon.gif
