+++ lib/pm_spotlight_shared/fifo_metadata.rb
